♻️ refactor [#161] `CardDetailModal`에서 comment section을 컴포넌트로 추출
